% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{execute_pipeline}
\alias{execute_pipeline}
\title{runs the FEM pipeline
provide classes argument as c("negativeclassname", "positiveclassname")}
\usage{
execute_pipeline(
  phenotype_file_name,
  read_count_dir_path,
  read_count_file_name,
  skip_row_count = 0,
  row_count = -1,
  na_strings = "NA",
  classification_criteria,
  filter_expression,
  classes,
  extracted_count_file_name = "read_counts.txt",
  output_label_file_name = "output_labels.txt",
  dataset_id,
  cores = 16,
  results_dir_path = "results"
)
}
\arguments{
\item{phenotype_file_name}{Name of the file containing phenotype info - class of each sample}
}
\description{
runs the FEM pipeline
provide classes argument as c("negativeclassname", "positiveclassname")
}
