% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction_arguments.R
\name{show_allowed_fems}
\alias{show_allowed_fems}
\title{show_allowed_fems}
\usage{
show_allowed_fems()
}
\description{
Gives list of allowed feature extraction methods that can be used in the pipeline
}
