% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{execute_pipeline}
\alias{execute_pipeline}
\title{FEM pipeline}
\usage{
execute_pipeline(
  phenotype_file_name,
  read_count_dir_path,
  read_count_file_name,
  sep = "",
  skip_row_count = 0,
  row_count = -1,
  na_strings = "NA",
  classification_criteria,
  filter_expression = expression(TRUE),
  classes,
  extracted_count_file_name = "read_counts.txt",
  output_label_file_name = "output_labels.txt",
  dataset_id,
  cores = 3,
  results_dir_path = "results",
  fems_to_run = c()
)
}
\arguments{
\item{phenotype_file_name}{Name of the file containing phenotype info - class of each sample.
\itemize{
\item should be a tab separated file
\item should have one field named "Sample" and contain sample names as in the read count file
}

example phenotype file contents given below :

Sample Age Gender Technology    GBMVsControl     GBMVsGlioma

s1     60    M    Microarray    GBM     GBM

s2     65    M   RNASeq        GBM     GBM

s3     59    F    RNASeq        Control NA

s4     55    F    Microarray    NA        Glioma

s5     50    M     RNASeq        Control NA}

\item{sep}{field separator character in read count file}

\item{classification_criteria}{Column in the phenotype file to perform classification on
Eg : GBMVsControl}

\item{filter_expression}{Filtering to be done on the samples based on a column in the phenotype file
Eg: Age > 55}

\item{classes}{Classes to be compared : c("negativeclassname", "positiveclassname")}

\item{dataset_id}{An ID for the data to be written in results}

\item{cores}{Number of cores to be used for parallel processing}

\item{fems_to_run}{Vector of names of FEMs to run  Eg: c("t-test", "mrmr30", "mrmr50"))

To see all allowed FEMs use show_allowed_fems()}
}
\description{
Runs the Feature Extraction Method Comparsion pipeline for the given data
}
